ㄹ파일 분리
